Feature: Refactor input type for timeout from int to float so we can use fractions of seconds

+++ src/browserist/browser/check_if/__main__.py
@@ -15,7 +15,7 @@ class CheckIfDriverMethods(DriverMethods):
     def __init__(self, browser_driver: BrowserDriver, settings: BrowserSettings) -> None:
         super().__init__(browser_driver, settings)
 
-    def contains_text(self, xpath: str, regex: str, ignore_case: bool = True, timeout: int = timeout.DEFAULT) -> bool:
+    def contains_text(self, xpath: str, regex: str, ignore_case: bool = True, timeout: float = timeout.DEFAULT) -> bool:
         """Check if element contains text. The condition works for both ordinary text (e.g. "Submit") or regular expression (e.g. r"colou?r"). Note it's a search for text, not a strict text match."""
 
         return check_if_contains_text(self._driver, xpath, regex, ignore_case, timeout)
@@ -45,7 +45,7 @@ class CheckIfDriverMethods(DriverMethods):
 
         return check_if_is_enabled(self._driver, xpath)
 
-    def is_image_loaded(self, xpath: str, timeout: int = timeout.DEFAULT) -> bool:
+    def is_image_loaded(self, xpath: str, timeout: float = timeout.DEFAULT) -> bool:
         """Check is image is loaded and ready in the DOM."""
 
         return check_if_is_image_loaded(self._driver, xpath, timeout)
