+++ src/browserist/browser/get/__main__.py
@@ -21,24 +21,24 @@ class GetDriverMethods(DriverMethods):
         self.attribute: GetAttributeDriverMethods = GetAttributeDriverMethods(browser_driver, settings)
         self.url: GetUrlDriverMethods = GetUrlDriverMethods(browser_driver, settings)
 
-    def dimensions(self, xpath: str, timeout: int = timeout.DEFAULT) -> tuple[int, int]:
+    def dimensions(self, xpath: str, timeout: float = timeout.DEFAULT) -> tuple[int, int]:
         """Get width and height of element in pixels. Usage:
 
         width, height = browser.get.dimensions("/element/xpath")"""
 
         return get_dimensions(self._driver, xpath, timeout)
 
-    def element(self, xpath: str, timeout: int = timeout.DEFAULT) -> object:
+    def element(self, xpath: str, timeout: float = timeout.DEFAULT) -> object:
         """Get single web element by XPath."""
 
         return get_element(self._driver, xpath, timeout)
 
-    def elements(self, xpath: str, timeout: int = timeout.DEFAULT) -> list[object]:
+    def elements(self, xpath: str, timeout: float = timeout.DEFAULT) -> list[object]:
         """Get multiple web elements by XPath."""
 
         return get_elements(self._driver, xpath, timeout)
 
-    def elements_by_tag(self, tag: str, timeout: int = timeout.DEFAULT) -> list[object]:
+    def elements_by_tag(self, tag: str, timeout: float = timeout.DEFAULT) -> list[object]:
         """"Get multiple web elements by HTML tag. Examples: "img" as tag for all <img> images, "a" for all <a> links, etc."""
 
         return get_elements_by_tag(self._driver, tag, timeout)
@@ -48,14 +48,14 @@ class GetDriverMethods(DriverMethods):
 
         return get_page_title(self._driver)
 
-    def text(self, xpath: str, timeout: int = timeout.DEFAULT) -> str:
+    def text(self, xpath: str, timeout: float = timeout.DEFAULT) -> str:
         """Get text from element.
 
         This method assumes that the text field shouldn't be empty and therefore will retry to get the text (for better support of single-page apps with extended loading time)."""
 
         return get_text(self._driver, xpath, timeout)
 
-    def texts(self, xpath: str, timeout: int = timeout.DEFAULT) -> list[str]:
+    def texts(self, xpath: str, timeout: float = timeout.DEFAULT) -> list[str]:
         """Get array of texts from elements.
 
         Assumes that the XPath targets multiple elements."""
