+++ src/browserist/browser/get/url/from_link.py
@@ -5,7 +5,7 @@ from ...wait.for_element import wait_for_element
 from ..attribute.value import get_attribute_value
 
 
-def get_url_from_link(driver: object, xpath: str, timeout: int = timeout.DEFAULT) -> str:
+def get_url_from_link(driver: object, xpath: str, timeout: float = timeout.DEFAULT) -> str:
     def get_href_attribute_of_element(driver: object, xpath: str) -> str:
         return get_attribute_value(driver, xpath, "href", constant.timeout.BYPASS)
