+++ src/browserist/browser/select/input_field.py
@@ -5,7 +5,7 @@ from ...model.type.xpath import XPath
 from ..wait.for_element import wait_for_element
 
 
-def select_input_field(driver: object, xpath: str, timeout: int = timeout.DEFAULT) -> None:
+def select_input_field(driver: object, xpath: str, timeout: float = timeout.DEFAULT) -> None:
     xpath = XPath(xpath)
     wait_for_element(driver, xpath, timeout)
     driver.find_element(By.XPATH, xpath).click()  # type: ignore
